# lcm/lcm/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "VNF Manager LCM Service rest API"
    },
    "basePath": "/gvnfmapi/lcm/v1",
...
    "paths": {
        "/vnf_instances": {
            "post": {
                "tags": [
                    "lcm Resource"
                ],
                "summary": "Create a VNF instance resource",
                "description": "Create a VNF instance resource",
                "operationId": "CreateVnf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The request params to create a VNF instance",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateVnfRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "A VNF Instance identifier was created successfully.",
                        "schema": {
                            "$ref": "#/definitions/CreateVnfResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Failed to process the request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "lcm Resource"
                ],
                "summary": "Query multiple VNF instances",
                "description": "Query multiple VNF instances",
                "operationId": "QueryMultipleVnf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "The request has succeeded.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/VnfInstanceInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Failed to process the request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/vnf_instances/{vnfInstanceId}": {
            "delete": {
                "tags": [
                    "lcm Resource"
                ],
                "summary": "Delete a VNF instance resource",
                "description": "Delete a VNF instance resource",
                "operationId": "DeleteVnf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "vnfInstanceId",
                        "description": "Identifier of the VNF instance.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The VNF instance resource and the associated VNF identifier were deleted successfully."
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "The VNF instance resource does not exist."
                    },
                    "500": {
                        "description": "Failed to process the request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "lcm Resource"
                ],
                "summary": "Query a VNF instance",
                "description": "Query a VNF instance",
                "operationId": "QuerySingleVnfs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "vnfInstanceId",
                        "description": "Identifier of the VNF instance.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request has succeeded.",
                        "schema": {
                            "$ref": "#/definitions/VnfInstanceInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "The VNF instance resource does not exist."
                    },
                    "500": {
                        "description": "Failed to process the request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/vnf_instances/{vnfInstanceId}/instantiate": {
            "post": {
                "tags": [
                    "lcm Resource"
                ],
                "summary": "Instantiate a VNF",
                "description": "Instantiate a VNF",
                "operationId": "InstantiateVnf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "vnfInstanceId",
                        "description": "Identifier of the VNF instance.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The request params to instantiate a VNF",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/InstantiateVnfRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "The request is accepted for processing, but the processing has not been completed.",
                        "schema": {
                            "$ref": "#/definitions/VnfLcOpIdResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "The VNF instance resource does not exist."
                    },
                    "500": {
                        "description": "Failed to process the request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/vnf_instances/{vnfInstanceId}/terminate": {
            "post": {
                "tags": [
                    "lcm Resource"
                ],
                "summary": "Terminate a VNF",
                "description": "Terminate a VNF",
                "operationId": "TerminateVnf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "vnfInstanceId",
                        "description": "Identifier of the VNF instance.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The request params to terminate a VNF",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TerminateVnfRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "The request is accepted for processing, but the processing has not been completed.",
                        "schema": {
                            "$ref": "#/definitions/VnfLcOpIdResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "The VNF instance resource does not exist."
                    },
                    "500": {
                        "description": "Failed to process the request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "CreateVnfRequest": {
            "type": "object",
            "required": [
                "vnfdId",
                "vnfInstanceName"
            ],
            "properties": {
                "vnfdId": {
                    "type": "string",
                    "description": "Identifier that identifies the VNFD which defines the VNF instance to be created."
                },
                "vnfInstanceName": {
                    "type": "string",
                    "description": "Human-readable name of the VNF instance to be created."
                },
                "vnfInstanceDescription": {
                    "type": "string",
                    "description": "Human-readable description of the VNF instance to be created."
                }
            }
        },
        "CreateVnfResponse": {
            "type": "object",
            "required": [
                "vnfInstanceId"
            ],
            "properties": {
                "vnfInstanceId": {
                    "type": "string",
                    "description": "VNF instance identifier just created."
                }
            }
        },
        "InstantiateVnfRequest": {
            "type": "object",
            "required": [
                "flavourId"
            ],
            "properties": {
                "flavourId": {
                    "type": "string",
                    "description": "Identifier of the VNF deployment flavour to be instantiated."
                },
                "instantiationLevelId": {
                    "type": "string",
                    "description": "Identifier of the instantiation level of the deployment flavour to be instantiated. If not present, the default instantiation level as declared in the VNFD is instantiated."
                },
                "extVirtualLinks": {
                    "items": {
                        "$ref": "#/definitions/ExtVirtualLinkData"
                    },
                    "description": "Information about external VLs to connect the VNF to."
                },
                "localizationLanguage": {
                    "type": "string",
                    "description": "Localization language of the VNF to be instantiated can be declared in the VNFD. The value shall comply with the format defined in IETF RFC 5646 [6]. "
                },
                "additionalParams": {
                    "type": "object",
                    "description": "Additional input parameters for the instantiation process, specific to the VNF being instantiated."
                }
            }
        },
        "TerminateVnfRequest": {
            "type": "object",
            "required": [
                "terminationType"
            ],
            "properties": {
                "terminationType": {
                    "type": "string",
                    "description": "Indicates whether forceful or graceful termination is requested.",
                    "enum": [
                        "FORCEFUL",
                        "GRACEFUL"
                    ]
                },
                "gracefulTerminationTimeout": {
                    "type": "integer",
                    "format": "int32",
					"minimum": "1",
                    "description": "This attribute is only applicable in case of graceful termination. It defines the time to wait for the VNF to be taken out of service before shutting down the VNF and releasing the resources. The unit is seconds. If not given and the \"terminationType\" attribute is set to \"GRACEFUL\", it is expected that the VNFM waits for the successful taking out of service of the VNF, no matter how long it takes, before shutting down the VNF and releasing the resources."
                }
            }
        },
        "VnfLcOpIdResponse": {
            "type": "object",
            "description": "VNF lifecycle operation occurrence identifier response",
            "required": [
                "vnfLcOpId"
            ],
            "properties": {
                "vnfLcOpId": {
                    "type": "string",
                    "description": "Identifier of a VNF lifecycle operation occurrence"
                }
            }
        },
        "ExtVirtualLinkData": {
            "type": "object",
            "required": [
                "resourceId",
                "extCps"
            ],
            "properties": {
                "extVirtualLinkId": {
                    "type": "string",
                    "description": "Identifier of the VL instance."
                },
                "vim": {
                    "$ref": "#/definitions/VimInfo",
                    "description": "Information about the VIM that manages this resource. This attribute shall be supported and present if VNF-related resource management in direct mode is applicable."
                },
                "resourceId": {
                    "type": "string",
                    "description": "The network UUID of VIM"
                },
                "extCps": {
                    "type": "array",
                    "description": "External CPs of the VNF to be connected to this external VL.",
                    "items": {
                        "$ref": "#/definitions/VnfExtCpData"
                    }
                }
            }
        },
        "VimInfo": {
            "type": "object",
            "required": [
                "vimInfoId",
                "vimId",
                "interfaceInfo",
                "accessInfo",
                "interfaceEndpoint"
            ],
            "properties": {
                "vimInfoId": {
                    "type": "string",
                    "description": "The identifier of this VimInfo instance, for the purpose of referencing it from other information elements."
                },
                "vimId": {
                    "type": "string",
                    "description": "The identifier of the VIM."
                },
                "interfaceInfo": {
                    "$ref": "#/definitions/InterfaceInfo",
                    "description": "Information about the interface to the VIM, including VIM provider type, API version, and protocol type."
                },
                "accessInfo": {
                    "$ref": "#/definitions/AccessInfo",
                    "description": "Authentication credentials for accessing the VIM. Examples may include those to support different authentication schemes, e.g., OAuth, Token, etc. "
                },
                "interfaceEndpoint": {
                    "type": "string",
                    "description": "Information about the interface endpoint. An example is a URL."
                }
            }
        },
...
        "VnfInstanceInfo": {
            "type": "object",
            "required": [
                "vnfInstanceId",
                "vnfInstanceName",
                "vnfInstanceDescription",
                "onboardedVnfPkgInfoId",
                "vnfdId",
                "vnfdVersion",
                "vnfSoftwareVersion",
                "vnfProvider",
                "vnfProductName",
                "instantiationState"
            ],
            "properties": {
                "vnfInstanceId": {
                    "type": "string",
                    "description": "VNF instance identifier."
                },
                "vnfInstanceName": {
                    "type": "string",
                    "description": "VNF instance name."
                },
                "vnfInstanceDescription": {
                    "type": "string",
                    "description": "Human-readable description of the VNF instance."
                },
                "onboardedVnfPkgInfoId": {
                    "type": "string",
                    "description": "Identifier of information held by the NFVO about the specific VNF Package on which the VNF is based. This identifier was allocated by the NFVO."
                },
                "vnfdId": {
                    "type": "string",
                    "description": "Identifier of the VNFD on which the VNF instance is based."
                },
                "vnfdVersion": {
                    "type": "string",
                    "description": "Identifies the version of the VNFD. The value is copied from the VNFD."
                },
                "vnfSoftwareVersion": {
                    "type": "string",
                    "description": "Software version of the VNF. The value is copied from the VNFD."
                },
                "vnfProvider": {
                    "type": "string",
                    "description": "Name of the person or company providing the VNF. The value is copied from the VNFD."
                },
                "vnfProductName": {
                    "type": "string",
                    "description": "Name to identify the VNF Product. Such as VNF Application Type. The value is copied from the VNFD."
                },
                "instantiationState": {
                    "type": "string",
                    "description": "The instantiation state of the VNF.",
                    "enum": [
                        "NOT_INSTANTIATED",
                        "INSTANTIATED"
                    ]
                },
                "vnfConfigurableProperties": {
                    "type": "object",
                    "description": "Current values of the configurable properties of the VNF instance. Configurable properties as declared in the VNFD."
                },
                "metadata": {
                    "type": "object",
                    "description": "Additional metadata describing the VNF instance. This attribute can be modified with the Modify VNF information operation."
                },
                "extensions": {
                    "type": "object",
                    "description": "VNF-specific attributes. This attribute can be modified with the Modify VNF information operation."
                },
                "instantiatedVnfInfo": {
                    "$ref": "#/definitions/InstantiatedVnfInfo",
                    "description": "Information specific to an instantiated VNF instance. This attribute shall be present if the instantiateState attribute value is INSTANTIATED."
                }
            }
        },
        "InstantiatedVnfInfo": {
            "type": "object",
            "required": [
                "flavourId",
                "vnfState"
            ],
            "properties": {
                "flavourId": {
                    "type": "string",
                    "description": "Identifier of the VNF deployment flavour to be instantiated."
                },
                "vnfState": {
                    "type": "string",
                    "description": "State of the VNF instance",
                    "enum": [
                        "STARTED",
                        "STOPPED"
                    ]
                },
                "scaleStatus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ScaleInfo"
                    },
                    "description": "Scale status of the VNF, one entry per aspect. Represents for every scaling aspect how \"big\" the VNF has been scaled w.r.t. that aspect. This attribute shall be present if the VNF supports scaling."
                },
                "extCpInfo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CpInfo"
                    },
                    "description": "Information about the external CPs exposed by the VNF instance."
                },
                "extVirtualLink": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ExtVirtualLinkInfo"
                    },
                    "description": "Information about the external VLs the VNF instance is connected to."
                },
                "monitoringParameters": {
                    "type": "object",
                    "description": "Active monitoring parameters."
                },
                "localizationLanguage": {
                    "type": "string",
                    "description": "Localization language of the VNF to be instantiated."
                },
                "vimInfo": {
                    "type": "array",
                    "description": "Information about VIM(s) managing resources for the VNF instance. This attribute shall be supported and present if VNF-related resource management in direct mode is applicable.",
                    "items": {
                        "$ref": "#/definitions/VimInfo"
                    }
                },
                "vnfcResourceInfo": {
                    "type": "array",
                    "description": "Information about the virtualised compute and storage resource(s) used by the VNFCs of the VNF instance. This attribute shall be supported and present if VNF-related resource management in direct mode is applicable.",
                    "items": {
                        "$ref": "#/definitions/VnfcResourceInfo"
                    }
                },
                "virtualLinkResourceInfo": {
                    "type": "array",
                    "description": "Information about the virtualised network resource(s) used by the VLs of the VNF instance. This attribute shall be supported and present if VNF-related resource management in direct mode is applicable. ",
                    "items": {
                        "$ref": "#/definitions/VirtualLinkResourceInfo"
                    }
                },
                "virtualStorageResourceInfo": {
                    "type": "array",
                    "description": "Information about the virtualised storage resource(s) used as storage for the VNF instance. This attribute shall be supported and present if VNF-related resource management in direct mode is applicable.",
                    "items": {
                        "$ref": "#/definitions/VirtualStorageResourceInfo"
                    }
                }
            }
        },
        "ScaleInfo": {
            "type": "object",
            "required": [
                "aspectId",
                "scaleLevel"
            ],
            "properties": {
                "aspectId": {
                    "type": "string",
                    "description": "Identifier of the scaling aspect."
                },
                "scaleLevel": {
                    "type": "integer",
                    "format": "int32",
                    "description": "Indicates the scale level. The minimum value shall be 0 and the maximum value shall be <= maxScaleLevel as described in the VNFD."
                }
            }
        },
        "CpInfo": {
            "type": "object",
            "required": [
                "cpInstanceId",
                "cpdId"
            ],
            "properties": {
                "cpInstanceId": {
                    "type": "string",
                    "description": "Identifier of the CP instance."
                },
                "cpdId": {
                    "type": "string",
                    "description": "Identifier of the CPD, in the VNFD."
                },
                "addresses": {
                    "type": "array",
                    "description": "List of network addresses that have been configured on the CP.",
                    "items": {
                        "$ref": "#/definitions/NetworkAddress"
                    }
                }
            }
        },
        "ExtVirtualLinkInfo": {
            "type": "object",
            "required": [
                "extVirtualLinkId",
                "resourceHandle"
            ],
            "properties": {
                "extVirtualLinkId": {
                    "type": "string",
                    "description": "Identifier of the external VL."
                },
                "resourceHandle": {
                    "$ref": "#/definitions/ResourceHandle",
                    "description": "Identifier of the resource realizing this VL."
                },
                "linkPorts": {
                    "type": "array",
                    "description": "Link ports of this VL.",
                    "items": {
                        "$ref": "#/definitions/VnfLinkPort"
                    }
                }
            }
        },
...
        "VnfcResourceInfo": {
            "type": "object",
            "required": [
                "vnfcInstanceId",
                "vduId",
                "computeResource",
                "storageResourceIds"
            ],
            "properties": {
                "vnfcInstanceId": {
                    "type": "string",
                    "description": "Identifier of this VNFC instance."
                },
                "vduId": {
                    "type": "string",
                    "description": "Reference to the applicable VDU information element in the VNFD."
                },
                "computeResource": {
                    "$ref": "#/definitions/ResourceHandle",
                    "description": "Reference to the VirtualCompute resource."
                },
                "storageResourceIds": {
                    "type": "array",
                    "description": "Reference(s) to the VirtualStorage resource(s).",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "VirtualStorageResourceInfo": {
            "type": "object",
            "required": [
                "virtualStorageInstanceId",
                "virtualStorageDescId",
                "storageResource"
            ],
            "properties": {
                "virtualStorageInstanceId": {
                    "type": "string",
                    "description": "Identifier of this virtual storage resource instance."
                },
                "virtualStorageDescId": {
                    "type": "string",
                    "description": "Identifier of the VirtualStorageDesc in the VNFD."
                },
                "storageResource": {
                    "$ref": "#/definitions/ResourceHandle",
                    "description": "Reference to the VirtualStorage resource."
                }
            }
        },
        "VirtualLinkResourceInfo": {
            "type": "object",
            "required": [
                "virtualLinkInstanceId",
                "virtualLinkDescId",
                "networkResource"
            ],
            "properties": {
                "virtualLinkInstanceId": {
                    "type": "string",
                    "description": "Identifier of this VL instance."
                },
                "virtualLinkDescId": {
                    "type": "string",
                    "description": "Identifier of the Virtual Link Descriptor (VLD) in the VNFD."
                },
                "networkResource": {
                    "$ref": "#/definitions/ResourceHandle",
                    "description": "Reference to the VirtualNetwork resource."
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}